+++ src/lib/utils.js
@@ -449,3 +449,11 @@ export const graphColors = [
 export const labels = ['A+', 'A', 'A-', 'B+', 'B', 'B-', 'C+', 'C', 'C-', 'D+', 'D', 'D-', 'F'];
 
 export const boilerExamsCourses = ["MA15800", "MA16010", "MA16100", "MA16200", "MA26100", "MA26200", "MA26500", "MA26600", "MA30300", "CS15900", "CS17700", "CS25100", "CHM11500", "ECON25200", "ECE20002", "PHYS17200"];
+
+
+// somewhat bandaid fix for missing decsriptions
+export const sanitizeDescription = (data) => {
+	if (data.description && data.description.startsWith("<a href=")) {
+		data.description = "No Description Available";
+	}
+};
